Split migrate command black-box test
