+++ cmd/migrate/migrate_internal_test.go
@@ -1,0 +1,56 @@
+package migrate
+
+// White-box testing required: renderSafetyReport and countRenderedStatements
+// are unexported formatting primitives whose edge cases are cheaper and clearer
+// to pin directly than through live database planning flows.
+
+import (
+	"bytes"
+	"encoding/json"
+	"testing"
+
+	qt "github.com/frankban/quicktest"
+
+	"github.com/stokaro/ptah/migration/safety"
+)
+
+func TestRenderSafetyReportJSON(t *testing.T) {
+	c := qt.New(t)
+
+	var out bytes.Buffer
+	assessments := []safety.StatementAssessment{{
+		Index:     1,
+		NodeType:  "sql",
+		Statement: "DROP TABLE users",
+		Severity:  safety.Destructive,
+		Reason:    "drops a table",
+	}}
+
+	err := renderSafetyReport(&out, "json", assessments)
+
+	c.Assert(err, qt.IsNil)
+	c.Assert(out.String(), qt.Not(qt.Contains), "Safety Assessment")
+
+	var report safety.Report
+	c.Assert(json.Unmarshal(out.Bytes(), &report), qt.IsNil)
+	c.Assert(report.Highest, qt.Equals, safety.Destructive)
+	c.Assert(report.Destructive, qt.IsTrue)
+	c.Assert(report.Assessments, qt.DeepEquals, assessments)
+}
+
+func TestCountRenderedStatementsCountsSQLStringNotBytes(t *testing.T) {
+	c := qt.New(t)
+
+	sql := "CREATE TABLE users (id INTEGER PRIMARY KEY);\nCREATE INDEX idx_users_id ON users(id);\n"
+
+	c.Assert(countRenderedStatements(sql), qt.Equals, 2)
+}
+
+func TestCountRenderedStatementsSkipsCommentOnlyStatements(t *testing.T) {
+	c := qt.New(t)
+
+	sql := "-- generated by ptah\n\nCREATE TABLE users (id INTEGER PRIMARY KEY);\n"
+
+	c.Assert(countRenderedStatements(sql), qt.Equals, 1)
+	c.Assert(countRenderedStatements("-- generated by ptah\n"), qt.Equals, 0)
+}
