+++ cmd/migrate/migrate_test.go
@@ -1,58 +1,15 @@
-package migrate
+package migrate_test
 
 import (
 	"bytes"
-	"encoding/json"
 	"os"
 	"path/filepath"
 	"testing"
 
 	qt "github.com/frankban/quicktest"
-	"github.com/spf13/cobra"
 
-	"github.com/stokaro/ptah/migration/safety"
+	"github.com/stokaro/ptah/cmd/migrate"
 )
-
-func TestRenderSafetyReportJSON(t *testing.T) {
-	c := qt.New(t)
-
-	var out bytes.Buffer
-	assessments := []safety.StatementAssessment{{
-		Index:     1,
-		NodeType:  "sql",
-		Statement: "DROP TABLE users",
-		Severity:  safety.Destructive,
-		Reason:    "drops a table",
-	}}
-
-	err := renderSafetyReport(&out, "json", assessments)
-
-	c.Assert(err, qt.IsNil)
-	c.Assert(out.String(), qt.Not(qt.Contains), "Safety Assessment")
-
-	var report safety.Report
-	c.Assert(json.Unmarshal(out.Bytes(), &report), qt.IsNil)
-	c.Assert(report.Highest, qt.Equals, safety.Destructive)
-	c.Assert(report.Destructive, qt.IsTrue)
-	c.Assert(report.Assessments, qt.DeepEquals, assessments)
-}
-
-func TestCountRenderedStatementsCountsSQLStringNotBytes(t *testing.T) {
-	c := qt.New(t)
-
-	sql := "CREATE TABLE users (id INTEGER PRIMARY KEY);\nCREATE INDEX idx_users_id ON users(id);\n"
-
-	c.Assert(countRenderedStatements(sql), qt.Equals, 2)
-}
-
-func TestCountRenderedStatementsSkipsCommentOnlyStatements(t *testing.T) {
-	c := qt.New(t)
-
-	sql := "-- generated by ptah\n\nCREATE TABLE users (id INTEGER PRIMARY KEY);\n"
-
-	c.Assert(countRenderedStatements(sql), qt.Equals, 1)
-	c.Assert(countRenderedStatements("-- generated by ptah\n"), qt.Equals, 0)
-}
 
 func TestMigratePlanTextOutputContainsSQLNotASTPointers(t *testing.T) {
 	c := qt.New(t)
@@ -70,7 +27,7 @@ type User struct {
 `), 0o600), qt.IsNil)
 
 	var out bytes.Buffer
-	cmd := newTestMigratePlanCommand(&out)
+	cmd := migrate.NewMigrateCommand()
 	cmd.SetOut(&out)
 	cmd.SetErr(&out)
 	cmd.SetArgs([]string{
@@ -84,19 +41,4 @@ type User struct {
 	c.Assert(out.String(), qt.Contains, `CREATE TABLE "users"`)
 	c.Assert(out.String(), qt.Not(qt.Contains), "[0x")
 	c.Assert(out.String(), qt.Not(qt.Contains), "&{")
-}
-
-func newTestMigratePlanCommand(out *bytes.Buffer) *cobra.Command {
-	opts := options{}
-	cmd := &cobra.Command{
-		Use:          "plan",
-		SilenceUsage: true,
-		RunE: func(cmd *cobra.Command, _ []string) error {
-			return migrateCommandWithOptions(cmd, &opts)
-		},
-	}
-	registerFlags(cmd, &opts)
-	cmd.SetOut(out)
-	cmd.SetErr(out)
-	return cmd
 }
